requirements.txt
nltk
spacy
scikit-learn
pandas
numpy
PyMuPDF
python-docx
typer
streamlit
tk
pyttsx3
SpeechRecognition
deep-translator
fpdf
